requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
geopy
rich
duckduckgo-search
streamlit
faiss-cpu
numpy

